WebCam FUN
The tutorial is going to go through how to make cool effects using canvas and your webcam. From the get go the code was a bit out of date, after some quick googleing I managed to figure out the modern way of doing things. A small victory but it felt big! then I tried to make it so the camera acts as a mirror but had some troubles.